moved helper functions from UI code to predict.py
