+++ predict.py
@@ -1,4 +1,5 @@
 import torch
+import streamlit as st
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn import Embedding
@@ -11,33 +12,40 @@ from src.data_functions import *
 from src.model_functions import *
 
 
-max_length  = 10
-embedding_layer, word2idx, idx2word = torch_pretrained_embedding()
+@st.cache_data
+def load_embeddings():
+    embedding_layer, word2idx, idx2word = torch_pretrained_embedding(load=True)
+    return embedding_layer, word2idx, idx2word
 
-model = GRUNet(input_dim=max_length, 
+
+@st.cache_resource
+def load_model(_embedding_layer):
+    model = GRUNet(input_dim=10, 
                 hidden_dim=128,
                 num_classes=30,
                 num_layers=2,
-                embedding_layer=embedding_layer,
+                embedding_layer=_embedding_layer,
                 embedding_dim=50,
                 batch_size=3)
 
-model.load_state_dict(torch.load("emoji_model.pt"))
-model.eval()
-def predict(sentence_str,
-            model=model,
-            max_len=max_length, 
-            embed_layer=embedding_layer,
-            word_to_idx=word2idx,
-            idx_to_word=idx2word):
+    model.load_state_dict(torch.load("emoji_model.pt", map_location=lambda storage, loc: storage))
+
+    return model
+
+
+def prepare_data(sentence_str, word_to_idx, max_length=10):
+    sentence_str = np.array([sentence_str])
+    sentence_str = sentences_to_indices(sentence_str, word_to_idx, max_length)
+    prepared_data = torch.from_numpy(sentence_str)
+    return prepared_data
+
+
+def predict(prepared_data, model):
     """
     Returns emoji classification prediction for a sample test sentence
     """
-    sentence_str = np.array([sentence_str])
-    sentence_str = sentences_to_indices(sentence_str, word2idx, max_length)
-    sentence_str = torch.from_numpy(sentence_str)
     with torch.no_grad():
-        predictions = model.forward(sentence_str, predict=True)
+        predictions = model.forward(prepared_data, predict=True)
         _, max_prediction = torch.max(predictions, dim=1)
     result = max_prediction.numpy()
