+++ ui.py
@@ -5,9 +5,7 @@ import torch
 import emoji
 import unicodeit
 from src.map_pred_to_emoji import *
-#from predict import *
-from src.data_functions import *
-from src.model_functions import *
+from predict import *
 
 
 
@@ -42,46 +40,6 @@ st.write("""
 
 
 ####################################### Model Inference #############################################
-
-
-@st.cache_data
-def load_embeddings():
-    embedding_layer, word2idx, idx2word = torch_pretrained_embedding(load=True)
-    return embedding_layer, word2idx, idx2word
-
-
-@st.cache_resource
-def load_model(_embedding_layer):
-    model = GRUNet(input_dim=10, 
-                hidden_dim=128,
-                num_classes=30,
-                num_layers=2,
-                embedding_layer=_embedding_layer,
-                embedding_dim=50,
-                batch_size=3)
-
-    model.load_state_dict(torch.load("emoji_model.pt", map_location=lambda storage, loc: storage))
-
-    return model
-
-
-def prepare_data(sentence_str, word_to_idx, max_length=10):
-    sentence_str = np.array([sentence_str])
-    sentence_str = sentences_to_indices(sentence_str, word_to_idx, max_length)
-    prepared_data = torch.from_numpy(sentence_str)
-    return prepared_data
-
-
-def predict(prepared_data, model):
-    """
-    Returns emoji classification prediction for a sample test sentence
-    """
-    with torch.no_grad():
-        predictions = model.forward(prepared_data, predict=True)
-        _, max_prediction = torch.max(predictions, dim=1)
-    result = max_prediction.numpy()
-
-    return result
 
 
 def main():
